requirements.txt
Django==5.2.5
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
python-docx==1.1.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
